Set global RAT-PAC2 compile flag to `-Wall -Werror` (#200)
* remove noexcepts to remove -wterminate

* -Wall -Werror compiles

* Address comments from @tannerbk

* comment out debug message

* Tensorflow compile coverage

* Fix std::any_cast usages + db loading

* Temporarily disable `-Werr` to ease transition

---------

+++ src/physics/src/PhysicsList.cc
@@ -74,7 +74,7 @@ void PhysicsList::EnableThermalNeutronScattering() {
   // Get the elastic scattering process used for neutrons
   G4HadronicProcess *n_elastic_process = nullptr;
   G4ProcessVector *proc_vec = n_definition->GetProcessManager()->GetProcessList();
-  for (int i = 0; i < proc_vec->size(); i++) {
+  for (size_t i = 0; i < proc_vec->size(); i++) {
     G4VProcess *proc = proc_vec->operator[](i);
     if (proc->GetProcessSubType() == fHadronElastic && proc->GetProcessType() == fHadronic) {
       n_elastic_process = dynamic_cast<G4HadronicProcess *>(proc);
